requirements.txt
flask
requests
google.cloud
